# insighthub-backend/src/openapi/swagger.yaml
openapi: 3.0.1
info:
  title: InsightHub API
  description: API for managing all the abilities.
  version: 1.0.0
...
paths:
...
  /post/{postId}:
    get:
      tags:
        - Posts
      summary: Retrieve all posts
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of posts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/PostNotFound'
    put:
      tags:
        - Posts
      security:
        - BearerAuth: [ ]
      summary: Update a post entirely - will replace the existing post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
      responses:
        '201':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PostNotFound'

    patch:
      tags:
        - Posts
      security:
        - BearerAuth: [ ]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequestPartial'
      responses:
        '201':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PostNotFound'

    delete:
      tags:
        - Posts
      security:
        - BearerAuth: [ ]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PostNotFound'

  # Comment Routes
  /comment:
    get:
      tags:
        - Comments
      summary: Retrieve all comments
      responses:
        '200':
          description: List of comments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'

    post:
      tags:
        - Comments
      security:
        - BearerAuth: []
      summary: Create a new comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/PostNotFound'
  /comment/{commentId}:
    put:
      tags:
        - Comments
      security:
        - BearerAuth: []
      summary: Update a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CommentNotFound'

    patch:
      tags:
        - Comments
      security:
        - BearerAuth: [ ]
      summary: Update a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequestPartial'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CommentNotFound'

    delete:
      tags:
        - Comments
      security:
        - BearerAuth: []
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Comment deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CommentNotFound'
  /comment/post/{postId}:
    get:
      tags:
        - Comments
      summary: Get comments for a specific post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of comments for the post
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/PostNotFound'

  # Authentication Routes
...
components:
  schemas:
    Post:
      type: object
      required:
        - id
        - owner
        - title
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: The post ID
        title:
          type: string
          description: The title of the post
        content:
          type: string
          description: The content of the post
        owner:
          type: string
          description: The ID of the user who owns the post
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the post was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the post was last updated
    PostRequestPartial:
      type: object
      required: []
      properties:
        title:
          type: string
          description: The title of the post
        content:
          type: string
          description: The content of the post
    PostRequest:
      allOf:
        - $ref: '#/components/schemas/PostRequestPartial'
        - type: object
          required:
            - title
            - content
    Comment:
      type: object
      required:
        - id
        - postId
        - owner
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: The unique identifier of the comment
        postId:
          type: string
          description: The ID of the post the comment belongs to
        owner:
          type: string
          description: The owner of the comment
        content:
          type: string
          description: The content of the comment
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the comment was created
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the comment was last updated
    CommentRequestPartial:
      type: object
      required: []
      properties:
        postId:
          type: string
          description: The ID of the post the comment should belong to
        content:
          type: string
          description: The content of the comment
    CommentRequest:
      allOf:
        - $ref: '#/components/schemas/CommentRequestPartial'
        - type: object
          required:
            - postId
            - content
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  responses:
    BadRequest:
      description: Bad request - Missing or bad required fields
    Unauthorized:
      description: Unauthorized - Missing token
    Forbidden:
      description: Forbidden - Invalid or expired token
    PostNotFound:
        description: Post not found
    CommentNotFound:
        description: Comment not found
...